+++ nextgen/config/tests/template_config_test.rs
@@ -122,6 +122,7 @@ variables:
                 *config.variables.get("bool").unwrap(),
                 TemplateVariable::Boolean(TemplateVariableBoolSetting {
                     default: false,
+                    order: None,
                     prompt: Some("prompt".into()),
                     required: Some(true)
                 })
@@ -166,6 +167,7 @@ variables:
                 *config.variables.get("num").unwrap(),
                 TemplateVariable::Number(TemplateVariableNumberSetting {
                     default: 123,
+                    order: None,
                     prompt: Some("prompt".into()),
                     required: Some(false)
                 })
@@ -208,6 +210,7 @@ variables:
                 *config.variables.get("str").unwrap(),
                 TemplateVariable::String(TemplateVariableStringSetting {
                     default: "abc".into(),
+                    order: None,
                     prompt: None,
                     required: None
                 })
@@ -257,6 +260,7 @@ variables:
                 TemplateVariable::Enum(TemplateVariableEnumSetting {
                     default: TemplateVariableEnumDefault::String("a".into()),
                     multiple: None,
+                    order: None,
                     prompt: Some("prompt".into()),
                     values: vec![
                         TemplateVariableEnumValue::String("a".into()),
@@ -299,6 +303,7 @@ variables:
                 TemplateVariable::Enum(TemplateVariableEnumSetting {
                     default: TemplateVariableEnumDefault::Vec(vec!["a".into(), "c".into()]),
                     multiple: Some(true),
+                    order: None,
                     prompt: Some("prompt".into()),
                     values: vec![
                         TemplateVariableEnumValue::String("a".into()),